requirements.txt
astropy==5.3.4
blobfile==2.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.0
fonttools==4.44.0
fsspec==2023.10.0
-e git+https://github.com/forever208/DDPM-IP.git@9f953651e7c1e3d7ef2038505e8343a9d2fb2945#egg=guided_diffusion
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
lpips==0.1.4
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.1
mpi4py @ file:///croot/mpi4py_1671223370575/work
mpmath==1.3.0
networkx==3.2.1
nmmn==1.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
protobuf==4.24.4
pycryptodomex==3.19.0
pyerfa==2.0.1.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
six==1.16.0
sympy==1.12
tensorboardX==2.6.2.2
torch==2.1.0
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.0.7
zipp==3.17.0
